+++ vscode-ext/glyph-middleware.js
@@ -1064,7 +1064,7 @@ class GlyphCompressor {
 
   _createSourceMap() {
     return {
-      version: '1.18.0',
+      version: '1.19.0',
       level: this.level,
       provider: this.provider,
       profile: this.providerProfile,